Adding hack to bypass time consuming nml gen
This is a hack to the mct buildnml to check if the modelio namelist file already exists. If it does, it changes the only thing necessary for our use case - the log file. Bypassing the normal generation process cuts the time necessary for our 80 instance generation down from 40 seconds to 0.6 seconds, which saves us a lot of core hours (the 40 seconds is 100% serial time while we are using thousands of cores; this saves about 10% of the job's total runtime in core hours for a very significant savings).